+++ addon/road.py
@@ -26,7 +26,6 @@ class road:
         self.geometry = geometry
         self.geometry_solver = geometry_solver
         self.params = {}
-        self.stencil = None
 
     def create_object_3d(self, context, params_input):
         '''
@@ -234,8 +233,7 @@ class road:
         '''
             Return the two connection points for a split road.
         '''
-        road_split_lane_idx = self.params['road_split_lane_idx']
-        t_cp_split = self.road_split_lane_idx_to_t(road_split_lane_idx)
+        t_cp_split = self.road_split_lane_idx_to_t()
         if road_split_type == 'start':
             t = 0
             cp_base = self.geometry.params['point_start']
@@ -252,12 +250,13 @@ class road:
         else:
             return cp_split, cp_base
 
-    def road_split_lane_idx_to_t(self, road_split_lane_idx):
+    def road_split_lane_idx_to_t(self):
         '''
             Convert index of first splitting lane to t coordinate of left/right
             side of the split lane border. Return 0 if there is no split.
         '''
         t_cp_split = 0
+        road_split_lane_idx = self.params['road_split_lane_idx']
         # Check if there really is a split
         if self.params['road_split_type'] != 'none':
             # Calculate lane ID from split index
@@ -267,20 +266,20 @@ class road:
                 lane_id_split = -1 * road_split_lane_idx
             # Calculate t coordinate of split connecting point
             for idx in range(abs(lane_id_split)):
-                    if lane_id_split > 0:
-                        # Do not add lanes with 0 width
-                        if not ((self.params['road_split_type'] == 'start' and
-                                 self.params['lanes_left_widths_change'][idx] == 'open') or
-                                (self.params['road_split_type'] == 'end' and \
-                                 self.params['lanes_left_widths_change'][idx] == 'close')):
-                            t_cp_split += self.params['lanes_left_widths'][idx]
-                    else:
-                        # Do not add lanes with 0 width
-                        if not ((self.params['road_split_type'] == 'start' and
-                                 self.params['lanes_right_widths_change'][idx] == 'open') or
-                                (self.params['road_split_type'] == 'end' and \
-                                 self.params['lanes_right_widths_change'][idx] == 'close')):
-                            t_cp_split -= self.params['lanes_right_widths'][idx]
+                if lane_id_split > 0:
+                    # Do not add lanes with 0 width
+                    if not ((self.params['road_split_type'] == 'start' and
+                                self.params['lanes_left_widths_change'][idx] == 'open') or
+                            (self.params['road_split_type'] == 'end' and \
+                                self.params['lanes_left_widths_change'][idx] == 'close')):
+                        t_cp_split += self.params['lanes_left_widths'][idx]
+                else:
+                    # Do not add lanes with 0 width
+                    if not ((self.params['road_split_type'] == 'start' and
+                                self.params['lanes_right_widths_change'][idx] == 'open') or
+                            (self.params['road_split_type'] == 'end' and \
+                                self.params['lanes_right_widths_change'][idx] == 'close')):
+                        t_cp_split -= self.params['lanes_right_widths'][idx]
         return t_cp_split
 
     def get_width_road_left(self, lanes):
@@ -600,8 +599,18 @@ class road:
                 strip_is_road_mark.append(False)
             elif lane.side == 'center':
                 if lane.road_mark_type != 'none':
-                    strip_to_lane.append(idx_lane)
-                    strip_is_road_mark.append(True)
+                    if lane.road_mark_type == 'solid' or \
+                        lane.road_mark_type == 'broken':
+                        strip_to_lane.append(idx_lane)
+                        strip_is_road_mark.append(True)
+                    else:
+                        # Double line
+                        strip_to_lane.append(idx_lane)
+                        strip_to_lane.append(idx_lane)
+                        strip_to_lane.append(idx_lane)
+                        strip_is_road_mark.append(True)
+                        strip_is_road_mark.append(False)
+                        strip_is_road_mark.append(True)
             else:
                 # lane.side == 'right'
                 strip_to_lane.append(idx_lane)
@@ -660,9 +669,7 @@ class road:
                         idx_face += 1
                     elif lanes[idx_lane].road_mark_type == 'solid_solid':
                         materials[material].append(idx_face)
-                        materials['asphalt'].append(idx_face + 1)
-                        materials[material].append(idx_face + 2)
-                        idx_face += 3
+                        idx_face += 1
             else:
                 if lanes[idx_lane].type == 'median':
                     materials['grass'].append(idx_face)